quick-and-dirty move of CLB deletion to cloud_client.py
